WIP: Disable key generation if not on primary
this didn't take any effect. However, I realized that I kept getting the same error: "CryptographicException: The key {c6497613-b53e-4dc7-81ff-76e7fbb4ea6e} was not found in the key ring." So, it could be that I need to clear the cookies as it may still contains the old header.